pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "network-hydraulic"
version = "0.1.0"
description = "Hydraulic calculation framework for pipe networks"
authors = [{ name = "Hydraulic Framework Team" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.6",
    "typer>=0.12",
    "ruamel.yaml>=0.18",
    "fluids>=1.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "ruff>=0.4",
    "mypy>=1.9"
]

[project.scripts]
network-hydraulic = "network_hydraulic.cli.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
packages = ["network_hydraulic"]
strict = true
